Refactored code from workshop_creator_gui.py and model.py into their own packages/modules.
